Added lib with mach.Request and common middlewares
The most significant piece of work in this commit is the implementation
of mach.Request. One of these is created for every request the server
receives. It is responsible for calling downstream apps as well as
coercing their return values to a promise that is then passed back
upstream.

I wanted to see how it felt to write some common middleware using this
implementation, so I ported several over from strata. These include:

- mach.basicAuth for doing HTTP Basic auth
- mach.commonLogger for Apache httpd-style logging of requests
- mach.contentType for automatically setting the Content-Type header
- mach.favicon for rejecting requests for /favicon.ico
- mach.file for serving static files
- mach.mapper for location and host-based routing
- mach.router for pattern-based routing
- mach.urlMap for constructing a mach.mapper from location/app pairs

I also made some small prototypes to experiment with.