Remove mach.sessionCookie, add mach.session
mach.session is a flexible middleware that can use any one of several
different backends for session storage. This change includes storage
in client-side cookies as well as server-side in memory or Redis. The
default is to use client-side cookies for backwards compatibility and
for convenience when in development/testing.